+++ Changelog.md
@@ -25,3 +25,8 @@ Fix: pip installing python client now also installs `requests` dependencies, whi
 -----
 
 * Python Client: Use tenacity for all requests
+
+Next
+----
+
+* Python Client: `Client` can now be pickeled again.